Use magic values and change run_time field type

+++ workflows/models/task.py
@@ -45,4 +45,4 @@ class TaskExecution(models.Model):
     status = models.CharField(max_length=1, choices=StatusChoices.choices, default=StatusChoices.PENDING, db_index=True)
     celery_task_id = models.CharField(max_length=50, null=True, blank=True)
     log = models.FileField(upload_to=task_execution_log_file_directory_path, null=True, blank=True)
-    run_time = models.IntegerField(blank=True, null=True)
+    run_time = models.FloatField(blank=True, null=True)
